fix hanging from missing model synchronization
afaict the `model` has to be unwrapped from accelerates distributed wrapping, to not wait for the other gpus forever.
this happens when not using ema.

also since validation only runs on the main process and gpu, the other gpus timeout (ncc) waiting for it.
synchronizing within the validation function is a okay-ish liveliness-check and prevents timeouts or having to increase them.